requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cloudevents==1.2.0
deprecation==2.1.0
flask[async]==2.0.3
functions-framework==3.0.0
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
packaging==21.3
pyparsing==3.0.7
requests==2.27.1
urllib3==1.26.8
watchdog==1.0.2
google-cloud-logging==3.0.0
discord~=1.7.3
firebase-admin==5.2.0